feat: 증상리스트 관리 커스텀 훅 구현[FE-19]
