+++ src/features/diagnosis/lib/chunkArray.ts
@@ -1,0 +1,5 @@
+export const chunkArray = (arr: string[], size: number) => {
+    return Array.from({ length: Math.ceil(arr.length / size) }, (_, i) =>
+      arr.slice(i * size, i * size + size)
+    );
+  };
